refactor(seeds): use a json file to hold breed names
