--- packages/backend/seeds/data/nomenclature.ts
@@ -1,1049 +1,0 @@
-export const breeds = [
-  {
-    name: 'AFFENPINSCHER'
-  },
-  {
-    name: 'AFGHAN HOUND'
-  },
-  {
-    name: "AÏDI (CHIEN DE MONTAGNE DE L'ATLAS)"
-  },
-  {
-    name: 'AIREDALE TERRIER'
-  },
-  {
-    name: 'AKITA'
-  },
-  {
-    name: 'ALASKAN MALAMUTE'
-  },
-  {
-    name: 'ALPENLÄNDISCHE DACHSBRACKE'
-  },
-  {
-    name: 'AMERICAN AKITA'
-  },
-  {
-    name: 'AMERICAN COCKER SPANIEL'
-  },
-  {
-    name: 'AMERICAN FOXHOUND'
-  },
-  {
-    name: 'AMERICAN STAFFORDSHIRE TERRIER'
-  },
-  {
-    name: 'AMERICAN WATER SPANIEL'
-  },
-  {
-    name: 'ANGLO-FRANÇAIS DE PETITE VENERIE'
-  },
-  {
-    name: 'ANJING KINTAMANI-BALI'
-  },
-  {
-    name: 'APPENZELLER SENNENHUND'
-  },
-  {
-    name: 'ARIEGEOIS'
-  },
-  {
-    name: 'AUSTRALIAN CATTLE DOG'
-  },
-  {
-    name: 'AUSTRALIAN KELPIE'
-  },
-  {
-    name: 'AUSTRALIAN SHEPHERD'
-  },
-  {
-    name: 'AUSTRALIAN SILKY TERRIER'
-  },
-  {
-    name: 'AUSTRALIAN STUMPY TAIL CATTLE DOG'
-  },
-  {
-    name: 'AUSTRALIAN TERRIER'
-  },
-  {
-    name: 'AZAWAKH'
-  },
-  {
-    name: 'BARBET'
-  },
-  {
-    name: 'BASENJI'
-  },
-  {
-    name: 'BASSET ARTESIEN NORMAND'
-  },
-  {
-    name: 'BASSET BLEU DE GASCOGNE'
-  },
-  {
-    name: 'BASSET FAUVE DE BRETAGNE'
-  },
-  {
-    name: 'BASSET HOUND'
-  },
-  {
-    name: 'BAYERISCHER GEBIRGSSCHWEISSHUND'
-  },
-  {
-    name: 'BEAGLE'
-  },
-  {
-    name: 'BEAGLE HARRIER'
-  },
-  {
-    name: 'BEARDED COLLIE'
-  },
-  {
-    name: 'BEDLINGTON TERRIER'
-  },
-  {
-    name: 'BERGER BLANC SUISSE'
-  },
-  {
-    name: 'BERGER DE BEAUCE'
-  },
-  {
-    name: 'BERGER DE BRIE'
-  },
-  {
-    name: 'BERGER PICARD'
-  },
-  {
-    name: 'BERNER SENNENHUND'
-  },
-  {
-    name: 'BICHON A POIL FRISE'
-  },
-  {
-    name: 'BICHON HAVANAIS'
-  },
-  {
-    name: 'BILLY'
-  },
-  {
-    name: 'BLACK AND TAN COONHOUND'
-  },
-  {
-    name: 'BOLOGNESE'
-  },
-  {
-    name: 'BORDER COLLIE'
-  },
-  {
-    name: 'BORDER TERRIER'
-  },
-  {
-    name: 'BOSANSKI OSTRODLAKI GONIC-BARAK'
-  },
-  {
-    name: 'BOSTON TERRIER'
-  },
-  {
-    name: 'BOULEDOGUE FRANÇAIS'
-  },
-  {
-    name: 'BOUVIER DES ARDENNES'
-  },
-  {
-    name: 'BOUVIER DES FLANDRES/VLAAMSE KOEHOND'
-  },
-  {
-    name: 'BRACCO ITALIANO'
-  },
-  {
-    name: 'BRANDLBRACKE - (VIERÄUGL)'
-  },
-  {
-    name: "BRAQUE D'AUVERGNE"
-  },
-  {
-    name: "BRAQUE DE L'ARIEGE"
-  },
-  {
-    name: 'BRAQUE DU BOURBONNAIS'
-  },
-  {
-    name: 'BRAQUE FRANÇAIS - TYPE GASCOGNE'
-  },
-  {
-    name: 'BRAQUE FRANÇAIS - TYPE PYRENEES'
-  },
-  {
-    name: 'BRAQUE SAINT-GERMAIN'
-  },
-  {
-    name: 'BRIQUET GRIFFON VENDEEN'
-  },
-  {
-    name: 'BROHOLMER'
-  },
-  {
-    name: 'BULL TERRIER'
-  },
-  {
-    name: 'BULLDOG'
-  },
-  {
-    name: 'BULLMASTIFF'
-  },
-  {
-    name: 'CA DE BESTIAR'
-  },
-  {
-    name: 'CAIRN TERRIER'
-  },
-  {
-    name: 'CANAAN DOG'
-  },
-  {
-    name: 'CANADIAN ESKIMO DOG'
-  },
-  {
-    name: 'CANE CORSO ITALIANO'
-  },
-  {
-    name: 'CANE DA PASTORE BERGAMASCO'
-  },
-  {
-    name: 'CANE DA PASTORE MAREMMANO-ABRUZZESE'
-  },
-  {
-    name: 'CANICHE'
-  },
-  {
-    name: 'CÃO DA SERRA DA ESTRELA'
-  },
-  {
-    name: 'CÃO DA SERRA DE AIRES'
-  },
-  {
-    name: 'CÃO DE AGUA PORTUGUÊS'
-  },
-  {
-    name: 'CÃO DE CASTRO LABOREIRO'
-  },
-  {
-    name: 'CÃO FILA DE SÃO MIGUEL'
-  },
-  {
-    name: 'CAVALIER KING CHARLES SPANIEL'
-  },
-  {
-    name: 'CESKOSLOVENSKÝ VLCIAK'
-  },
-  {
-    name: 'CESKÝ FOUSEK'
-  },
-  {
-    name: 'CESKÝ TERIER'
-  },
-  {
-    name: 'CHART POLSKI'
-  },
-  {
-    name: 'CHESAPEAKE BAY RETRIEVER'
-  },
-  {
-    name: "CHIEN D'ARTOIS"
-  },
-  {
-    name: 'CHIEN DE BERGER BELGE'
-  },
-  {
-    name: 'CHIEN DE BERGER DES PYRENEES A FACE RASE'
-  },
-  {
-    name: 'CHIEN DE BERGER DES PYRENEES A POIL LONG'
-  },
-  {
-    name: 'CHIEN DE MONTAGNE DES PYRENEES'
-  },
-  {
-    name: 'CHIEN DE SAINT HUBERT'
-  },
-  {
-    name: 'CHIHUAHUEÑO'
-  },
-  {
-    name: 'CHIN'
-  },
-  {
-    name: 'CHINESE CRESTED DOG'
-  },
-  {
-    name: 'CHODSKY PES'
-  },
-  {
-    name: 'CHOW CHOW'
-  },
-  {
-    name: 'CIMARRÓN URUGUAYO'
-  },
-  {
-    name: 'CIOBANESC ROMÂNESC CARPATIN'
-  },
-  {
-    name: 'CIOBANESC ROMÂNESC DE BUCOVINA'
-  },
-  {
-    name: 'CIOBANESC ROMÂNESC MIORITIC'
-  },
-  {
-    name: "CIRNECO DELL'ETNA"
-  },
-  {
-    name: 'CLUMBER SPANIEL'
-  },
-  {
-    name: 'COLLIE ROUGH'
-  },
-  {
-    name: 'COLLIE SMOOTH'
-  },
-  {
-    name: 'COTON DE TULEAR'
-  },
-  {
-    name: 'CRNOGORSKI PLANINSKI GONIC'
-  },
-  {
-    name: 'CURLY COATED RETRIEVER'
-  },
-  {
-    name: 'DACHSHUND'
-  },
-  {
-    name: 'DALMATINSKI PAS'
-  },
-  {
-    name: 'DANDIE DINMONT TERRIER'
-  },
-  {
-    name: 'DANSK-SVENSK GÅRDSHUND'
-  },
-  {
-    name: 'DEERHOUND'
-  },
-  {
-    name: 'DEUTSCH DRAHTHAAR'
-  },
-  {
-    name: 'DEUTSCH KURZHAAR'
-  },
-  {
-    name: 'DEUTSCH LANGHAAR'
-  },
-  {
-    name: 'DEUTSCH STICHELHAAR'
-  },
-  {
-    name: 'DEUTSCHE BRACKE'
-  },
-  {
-    name: 'DEUTSCHE DOGGE'
-  },
-  {
-    name: 'DEUTSCHER BOXER'
-  },
-  {
-    name: 'DEUTSCHER JAGDTERRIER'
-  },
-  {
-    name: 'DEUTSCHER PINSCHER'
-  },
-  {
-    name: 'DEUTSCHER SCHÄFERHUND'
-  },
-  {
-    name: 'DEUTSCHER SPITZ'
-  },
-  {
-    name: 'DEUTSCHER WACHTELHUND'
-  },
-  {
-    name: 'DOBERMANN'
-  },
-  {
-    name: 'DOGO ARGENTINO'
-  },
-  {
-    name: 'DOGUE DE BORDEAUX'
-  },
-  {
-    name: 'DO-KHYI'
-  },
-  {
-    name: 'DRENTSCHE PATRIJSHOND'
-  },
-  {
-    name: 'DREVER'
-  },
-  {
-    name: 'DRÖTZÖRÜ MAGYAR VIZSLA'
-  },
-  {
-    name: 'DUNKER'
-  },
-  {
-    name: 'ENGLISH COCKER SPANIEL'
-  },
-  {
-    name: 'ENGLISH FOXHOUND'
-  },
-  {
-    name: 'ENGLISH POINTER'
-  },
-  {
-    name: 'ENGLISH SETTER'
-  },
-  {
-    name: 'ENGLISH SPRINGER SPANIEL'
-  },
-  {
-    name: 'ENGLISH TOY TERRIER'
-  },
-  {
-    name: 'ENTLEBUCHER SENNENHUND'
-  },
-  {
-    name: 'EPAGNEUL BLEU DE PICARDIE'
-  },
-  {
-    name: 'EPAGNEUL BRETON'
-  },
-  {
-    name: 'EPAGNEUL DE PONT-AUDEMER'
-  },
-  {
-    name: 'EPAGNEUL FRANÇAIS'
-  },
-  {
-    name: 'EPAGNEUL NAIN CONTINENTAL'
-  },
-  {
-    name: 'EPAGNEUL PICARD'
-  },
-  {
-    name: 'ERDÉLYI KOPÓ'
-  },
-  {
-    name: 'EURASIER'
-  },
-  {
-    name: 'FIELD SPANIEL'
-  },
-  {
-    name: 'FILA BRASILEIRO'
-  },
-  {
-    name: 'FLAT COATED RETRIEVER'
-  },
-  {
-    name: 'FOX TERRIER (SMOOTH)'
-  },
-  {
-    name: 'FOX TERRIER (WIRE)'
-  },
-  {
-    name: 'FRANÇAIS BLANC ET NOIR'
-  },
-  {
-    name: 'FRANÇAIS BLANC ET ORANGE'
-  },
-  {
-    name: 'FRANÇAIS TRICOLORE'
-  },
-  {
-    name: 'GALGO ESPAÑOL'
-  },
-  {
-    name: 'GAMMEL DANSK HØNSEHUND'
-  },
-  {
-    name: 'GASCON SAINTONGEOIS'
-  },
-  {
-    name: 'GOLDEN RETRIEVER'
-  },
-  {
-    name: 'GONCZY POLSKI'
-  },
-  {
-    name: 'GORDON SETTER'
-  },
-  {
-    name: "GOS D'ATURA CATALÁ"
-  },
-  {
-    name: 'GRAND ANGLO-FRANÇAIS BLANC ET NOIR'
-  },
-  {
-    name: 'GRAND ANGLO-FRANÇAIS BLANC ET ORANGE'
-  },
-  {
-    name: 'GRAND ANGLO-FRANÇAIS TRICOLORE'
-  },
-  {
-    name: 'GRAND BASSET GRIFFON VENDEEN'
-  },
-  {
-    name: 'GRAND BLEU DE GASCOGNE'
-  },
-  {
-    name: 'GRAND GRIFFON VENDEEN'
-  },
-  {
-    name: 'GREYHOUND'
-  },
-  {
-    name: 'GRIFFON A POIL DUR KORTHALS'
-  },
-  {
-    name: 'GRIFFON BELGE'
-  },
-  {
-    name: 'GRIFFON BLEU DE GASCOGNE'
-  },
-  {
-    name: 'GRIFFON BRUXELLOIS'
-  },
-  {
-    name: 'GRIFFON FAUVE DE BRETAGNE'
-  },
-  {
-    name: 'GRIFFON NIVERNAIS'
-  },
-  {
-    name: 'GRØNLANDSHUND'
-  },
-  {
-    name: 'GROSSER MÜNSTERLÄNDER VORSTEHHUND'
-  },
-  {
-    name: 'GROSSER SCHWEIZER SENNENHUND'
-  },
-  {
-    name: 'HALDENSTØVER'
-  },
-  {
-    name: 'HAMILTONSTÖVARE'
-  },
-  {
-    name: 'HANNOVERSCHER SCHWEISSHUND'
-  },
-  {
-    name: 'HARRIER'
-  },
-  {
-    name: 'HELLINIKOS ICHNILATIS'
-  },
-  {
-    name: 'HOKKAIDO'
-  },
-  {
-    name: 'HOLLANDSE HERDERSHOND'
-  },
-  {
-    name: 'HOLLANDSE SMOUSHOND'
-  },
-  {
-    name: 'HOVAWART'
-  },
-  {
-    name: 'HRVATSKI OVCAR'
-  },
-  {
-    name: 'HYGENHUND'
-  },
-  {
-    name: 'IRISH GLEN OF IMAAL TERRIER'
-  },
-  {
-    name: 'IRISH RED AND WHITE SETTER'
-  },
-  {
-    name: 'IRISH RED SETTER'
-  },
-  {
-    name: 'IRISH SOFT COATED WHEATEN TERRIER'
-  },
-  {
-    name: 'IRISH TERRIER'
-  },
-  {
-    name: 'IRISH WATER SPANIEL'
-  },
-  {
-    name: 'IRISH WOLFHOUND'
-  },
-  {
-    name: 'ISTARSKI OSTRODLAKI GONIC'
-  },
-  {
-    name: 'ISTARSKI KRATKODLAKI GONIC'
-  },
-  {
-    name: 'ÍSLENSKUR FJÁRHUNDUR'
-  },
-  {
-    name: 'JACK RUSSELL TERRIER'
-  },
-  {
-    name: 'JÄMTHUND'
-  },
-  {
-    name: 'JUGOSLOVENSKI OVCARSKI PAS-SARPLANINAC'
-  },
-  {
-    name: 'KAI'
-  },
-  {
-    name: 'KANGAL ÇOBAN KÖPEGI'
-  },
-  {
-    name: 'KARJALANKARHUKOIRA'
-  },
-  {
-    name: 'KAVKAZSKAÏA OVTCHARKA'
-  },
-  {
-    name: 'KERRY BLUE TERRIER'
-  },
-  {
-    name: 'KING CHARLES SPANIEL'
-  },
-  {
-    name: 'KISHU'
-  },
-  {
-    name: 'KLEINER MÜNSTERLÄNDER'
-  },
-  {
-    name: 'KOMONDOR'
-  },
-  {
-    name: 'KOREA JINDO DOG'
-  },
-  {
-    name: 'KRASKI OVCAR'
-  },
-  {
-    name: 'KROMFOHRLÄNDER'
-  },
-  {
-    name: 'KUVASZ'
-  },
-  {
-    name: 'LABRADOR RETRIEVER'
-  },
-  {
-    name: 'LAGOTTO ROMAGNOLO'
-  },
-  {
-    name: 'LAKELAND TERRIER'
-  },
-  {
-    name: 'LANCASHIRE HEELER'
-  },
-  {
-    name: 'LANDSEER (EUROPÄISCH-KONTINENTALER TYP)'
-  },
-  {
-    name: 'LAPINPOROKOIRA'
-  },
-  {
-    name: 'LEONBERGER'
-  },
-  {
-    name: 'LHASA APSO'
-  },
-  {
-    name: 'MAGYAR AGAR'
-  },
-  {
-    name: 'MALTESE'
-  },
-  {
-    name: 'MANCHESTER TERRIER'
-  },
-  {
-    name: 'MASTIFF'
-  },
-  {
-    name: 'MASTÍN DEL PIRINEO'
-  },
-  {
-    name: 'MASTÍN ESPAÑOL'
-  },
-  {
-    name: 'MASTINO NAPOLETANO'
-  },
-  {
-    name: 'MINIATURE BULL TERRIER'
-  },
-  {
-    name: 'MUDI'
-  },
-  {
-    name: 'NEDERLANDSE KOOIKERHONDJE'
-  },
-  {
-    name: 'NEDERLANDSE SCHAPENDOES'
-  },
-  {
-    name: 'NEWFOUNDLAND'
-  },
-  {
-    name: 'NIHON SUPITTSU'
-  },
-  {
-    name: 'NIHON TERIA'
-  },
-  {
-    name: 'NORFOLK TERRIER'
-  },
-  {
-    name: 'NORRBOTTENSPETS'
-  },
-  {
-    name: 'NORSK BUHUND'
-  },
-  {
-    name: 'NORSK ELGHUND GRÅ'
-  },
-  {
-    name: 'NORSK ELGHUND SORT'
-  },
-  {
-    name: 'NORSK LUNDEHUND'
-  },
-  {
-    name: 'NORWICH TERRIER'
-  },
-  {
-    name: 'NOVA SCOTIA DUCK TOLLING RETRIEVER'
-  },
-  {
-    name: 'OGAR POLSKI'
-  },
-  {
-    name: 'OLD ENGLISH SHEEPDOG'
-  },
-  {
-    name: 'OTTERHOUND'
-  },
-  {
-    name: 'ÖSTERREICHISCHER PINSCHER'
-  },
-  {
-    name: 'PARSON RUSSELL TERRIER'
-  },
-  {
-    name: 'PEKINGESE'
-  },
-  {
-    name: 'PERDIGUEIRO PORTUGUÊS'
-  },
-  {
-    name: 'PERDIGUERO DE BURGOS'
-  },
-  {
-    name: 'PERRO DE AGUA ESPAÑOL'
-  },
-  {
-    name: 'PERRO DOGO MALLORQUÍN'
-  },
-  {
-    name: 'PERRO SIN PELO DEL PERÚ'
-  },
-  {
-    name: 'PETIT BASSET GRIFFON VENDEEN'
-  },
-  {
-    name: 'PETIT BLEU DE GASCOGNE'
-  },
-  {
-    name: 'PETIT BRABANÇON'
-  },
-  {
-    name: 'PETIT CHIEN LION'
-  },
-  {
-    name: 'PHARAOH HOUND'
-  },
-  {
-    name: 'PICCOLO LEVRIERO ITALIANO'
-  },
-  {
-    name: 'PODENCO CANARIO'
-  },
-  {
-    name: 'PODENCO IBICENCO'
-  },
-  {
-    name: 'PODENGO PORTUGUÊS'
-  },
-  {
-    name: 'POITEVIN'
-  },
-  {
-    name: 'POLSKI OWCZAREK NIZINNY'
-  },
-  {
-    name: 'POLSKI OWCZAREK PODHALANSKI'
-  },
-  {
-    name: 'PORCELAINE'
-  },
-  {
-    name: 'POSAVSKI GONIC'
-  },
-  {
-    name: 'PRAZSKY KRYSARIK'
-  },
-  {
-    name: 'PRESA CANARIO'
-  },
-  {
-    name: 'PUDELPOINTER'
-  },
-  {
-    name: 'PUG'
-  },
-  {
-    name: 'PULI'
-  },
-  {
-    name: 'PUMI'
-  },
-  {
-    name: 'RAFEIRO DO ALENTEJO'
-  },
-  {
-    name: 'RHODESIAN RIDGEBACK'
-  },
-  {
-    name: 'RIESENSCHNAUZER'
-  },
-  {
-    name: 'ROTTWEILER'
-  },
-  {
-    name: 'RÖVIDSZÖRÜ MAGYAR VIZSLA'
-  },
-  {
-    name: 'RUSSKAYA PSOVAYA BORZAYA'
-  },
-  {
-    name: 'RUSSKIY TCHIORNY TERRIER'
-  },
-  {
-    name: 'RUSSKIY TOY'
-  },
-  {
-    name: 'RUSSKO-EVROPEÏSKAÏA LAÏKA'
-  },
-  {
-    name: 'SAARLOOSWOLFHOND'
-  },
-  {
-    name: 'SABUESO ESPAÑOL'
-  },
-  {
-    name: 'SALUKI'
-  },
-  {
-    name: 'SAMOIEDSKAÏA SABAKA'
-  },
-  {
-    name: 'SCHILLERSTÖVARE'
-  },
-  {
-    name: 'SCHIPPERKE'
-  },
-  {
-    name: 'SCHNAUZER'
-  },
-  {
-    name: 'SCHWEIZER NIEDERLAUFHUND'
-  },
-  {
-    name: 'SCHWEIZER LAUFHUND - CHIEN COURANT SUISSE'
-  },
-  {
-    name: 'SCOTTISH TERRIER'
-  },
-  {
-    name: 'SEALYHAM TERRIER'
-  },
-  {
-    name: 'SEGUGIO ITALIANO A PELO FORTE'
-  },
-  {
-    name: 'SEGUGIO ITALIANO A PELO RASO'
-  },
-  {
-    name: 'SEGUGIO MAREMMANO'
-  },
-  {
-    name: 'SHAR PEI'
-  },
-  {
-    name: 'SHETLAND SHEEPDOG'
-  },
-  {
-    name: 'SHIBA'
-  },
-  {
-    name: 'SHIH TZU'
-  },
-  {
-    name: 'SHIKOKU'
-  },
-  {
-    name: 'SIBERIAN HUSKY'
-  },
-  {
-    name: 'SKYE TERRIER'
-  },
-  {
-    name: 'SLOUGHI'
-  },
-  {
-    name: 'SLOVENSKÝ CUVAC'
-  },
-  {
-    name: 'SLOVENSKÝ HRUBOSRSTÝ STAVAC (OHAR)'
-  },
-  {
-    name: 'SLOVENSKÝ KOPOV'
-  },
-  {
-    name: 'SMÅLANDSSTÖVARE'
-  },
-  {
-    name: 'SPINONE ITALIANO'
-  },
-  {
-    name: 'SREDNEASIATSKAYA OVTCHARKA'
-  },
-  {
-    name: 'SRPSKI GONIC'
-  },
-  {
-    name: 'SRPSKI TROBOJNI GONIC'
-  },
-  {
-    name: 'ST.BERNHARDSHUND - BERNHARDINER'
-  },
-  {
-    name: 'STABIJHOUN'
-  },
-  {
-    name: 'STAFFORDSHIRE BULL TERRIER'
-  },
-  {
-    name: 'STEIRISCHE RAUHHAARBRACKE'
-  },
-  {
-    name: 'SUOMENAJOKOIRA'
-  },
-  {
-    name: 'SUOMENLAPINKOIRA'
-  },
-  {
-    name: 'SUOMENPYSTYKORVA'
-  },
-  {
-    name: 'SUSSEX SPANIEL'
-  },
-  {
-    name: 'SVENSK LAPPHUND'
-  },
-  {
-    name: 'TAIWAN DOG'
-  },
-  {
-    name: 'TERRIER BRASILEIRO'
-  },
-  {
-    name: 'THAI BANGKAEW DOG'
-  },
-  {
-    name: 'THAI RIDGEBACK DOG'
-  },
-  {
-    name: 'TIBETAN SPANIEL'
-  },
-  {
-    name: 'TIBETAN TERRIER'
-  },
-  {
-    name: 'TIROLER BRACKE'
-  },
-  {
-    name: 'TORNJAK'
-  },
-  {
-    name: 'TOSA'
-  },
-  {
-    name: 'VÄSTGÖTASPETS'
-  },
-  {
-    name: 'VOLPINO ITALIANO'
-  },
-  {
-    name: 'VOSTOTCHNO-SIBIRSKAÏA LAÏKA'
-  },
-  {
-    name: 'WEIMARANER'
-  },
-  {
-    name: 'WELSH CORGI (CARDIGAN)'
-  },
-  {
-    name: 'WELSH CORGI (PEMBROKE)'
-  },
-  {
-    name: 'WELSH SPRINGER SPANIEL'
-  },
-  {
-    name: 'WELSH TERRIER'
-  },
-  {
-    name: 'WEST HIGHLAND WHITE TERRIER'
-  },
-  {
-    name: 'WESTFÄLISCHE DACHSBRACKE'
-  },
-  {
-    name: 'WETTERHOUN'
-  },
-  {
-    name: 'WHIPPET'
-  },
-  {
-    name: 'XOLOITZCUINTLE'
-  },
-  {
-    name: 'YORKSHIRE TERRIER'
-  },
-  {
-    name: 'YUZHNORUSSKAYA OVCHARKA'
-  },
-  {
-    name: 'ZAPADNO-SIBIRSKAÏA LAÏKA'
-  },
-  {
-    name: 'ZWERGPINSCHER'
-  },
-  {
-    name: 'ZWERGSCHNAUZER'
-  }
-];
