+++ tsconfig.json
@@ -12,7 +12,7 @@
     "allowJs": true,
     "noImplicitAny": false,
     "esModuleInterop": true,
-
+    "resolveJsonModule": false,
     "allowSyntheticDefaultImports": true,
     "experimentalDecorators": true
   },
